requirements.txt
streamlit>=1.30.0
pandas>=2.0.0
openpyxl>=3.1.0
plotly>=5.18.0
yfinance>=0.2.0
python-dateutil>=2.8.2
google-generativeai>=0.3.0
anthropic>=0.30.0
alpaca-py>=0.30.0
pandas-datareader>=0.10.0
python-dotenv>=1.0.0
gspread>=6.0.0
google-auth>=2.0.0
setuptools>=70.0.0
